refactor: extract components to eliminate lint warnings
Eliminated all no-magic-numbers, no-multi-comp, no-negated-condition,
and curly warnings by extracting shared components and using named
constants for time thresholds.

Extracted: DashboardShell, FilterBar, FilteredStoreGrid,
SectionHeaderButton, CollapsedHint, MenuItem, useClickOutside hook.

Remaining 10 max-lines-per-function warnings are inherent to React
JSX components exceeding the 25-line limit.
