+++ app/frontend/pages/admin/dashboard.tsx
@@ -1,11 +1,9 @@
 import { useState } from "react";
 
-import MilkcrateShell from "@/layouts/milkcrate_shell";
 import type { AdminDashboardProps } from "@/types/inertia";
 
-import { DashboardHeader } from "./dashboard/dashboard_header";
 import { DashboardPanels } from "./dashboard/dashboard_panels";
-import { FlashBannerIfNotice } from "./dashboard/flash_banner_if_notice";
+import { DashboardShell } from "./dashboard/dashboard_shell";
 import { useResync } from "./dashboard/use_resync";
 
 export type { HealthFilter, HealthFilterKey } from "./dashboard/dashboard_constants";
@@ -21,26 +19,19 @@ export default function Dashboard({
   const [healthFilter, setHealthFilter] = useState<HealthFilter>([]);
 
   return (
-    <div className="min-h-screen bg-mc-bg text-mc-text">
-      <MilkcrateShell
-        header={
-          <DashboardHeader
-            active_stores={active_stores}
-            healthFilter={healthFilter}
-            onHealthFilterChange={setHealthFilter}
-          />
-        }
-        afterHeader={<FlashBannerIfNotice notice={notice} alert={alert} />}
-        contentWidth="max-w-7xl"
-        contentPadding="px-4 py-6 sm:px-6 lg:px-8"
-      >
-        <DashboardPanels
-          active_stores={active_stores}
-          healthFilter={healthFilter}
-          applicants={applicants}
-          discogs_onboarding={discogs_onboarding}
-        />
-      </MilkcrateShell>
-    </div>
+    <DashboardShell
+      active_stores={active_stores}
+      healthFilter={healthFilter}
+      onHealthFilterChange={setHealthFilter}
+      notice={notice}
+      alert={alert}
+    >
+      <DashboardPanels
+        active_stores={active_stores}
+        healthFilter={healthFilter}
+        applicants={applicants}
+        discogs_onboarding={discogs_onboarding}
+      />
+    </DashboardShell>
   );
 }
