+++ app/frontend/pages/admin/dashboard/collapsed_hint.tsx
@@ -1,0 +1,7 @@
+export function CollapsedHint({ count }: { count: number }) {
+  return (
+    <p className="text-sm text-mc-text-dim">
+      {count} store{count === 1 ? "" : "s"} hidden
+    </p>
+  );
+}
