+++ app/frontend/pages/admin/dashboard/section_content.tsx
@@ -15,11 +15,3 @@ export function ExpandedContent({ stores }: { stores: AdminStoreSummary[] }) {
   }
   return <EmptyState>No stores in this section.</EmptyState>;
 }
-
-export function CollapsedHint({ count }: { count: number }) {
-  return (
-    <p className="text-sm text-mc-text-dim">
-      {count} store{count === 1 ? "" : "s"} hidden
-    </p>
-  );
-}
