+++ app/frontend/pages/admin/dashboard/store_card/action_menu.tsx
@@ -1,25 +1,13 @@
-import { useEffect, useRef, useState } from "react";
+import { useState } from "react";
 
 import type { AdminStoreSummary } from "@/types/inertia";
 
 import { MenuItems } from "./menu_items";
+import { useClickOutside } from "./use_click_outside";
 
 export function ActionMenu({ store }: { store: AdminStoreSummary }) {
   const [open, setOpen] = useState(false);
-  const menuRef = useRef<HTMLDivElement>(null);
-
-  useEffect(() => {
-    if (!open) {
-      return;
-    }
-    const handleClick = (e: MouseEvent) => {
-      if (menuRef.current && !menuRef.current.contains(e.target as Node)) {
-        setOpen(false);
-      }
-    };
-    document.addEventListener("mousedown", handleClick);
-    return () => document.removeEventListener("mousedown", handleClick);
-  }, [open]);
+  const menuRef = useClickOutside(() => setOpen(false));
 
   return (
     <div ref={menuRef} className="relative">
