+++ app/frontend/pages/admin/dashboard/store_card/use_click_outside.ts
@@ -1,0 +1,17 @@
+import { useEffect, useRef } from "react";
+
+export function useClickOutside(onClose: () => void) {
+  const ref = useRef<HTMLDivElement>(null);
+
+  useEffect(() => {
+    const handleClick = (e: MouseEvent) => {
+      if (ref.current && !ref.current.contains(e.target as Node)) {
+        onClose();
+      }
+    };
+    document.addEventListener("mousedown", handleClick);
+    return () => document.removeEventListener("mousedown", handleClick);
+  }, [onClose]);
+
+  return ref;
+}
